Crate libsoxr

libsoxr-rs

This library is a thin wrapper for libsoxr which is a "High quality, one-dimensional sample-rate conversion library".

Use the soxr namespace to use soxr in your Rust program. For direct access to the libsoxr functions, you can use the api namespace.

Reexports

pub use soxr::Soxr;

pub use datatype::Datatype;

pub use spec::{IOSpec, QualitySpec, RuntimeSpec};

Modules

api
This is a thin wrapper around the libsoxr C API.

datatype
Encapsulated data types for samples

soxr
Rust API for SOXR.

spec
For specifying the runtime settings of the resampler For specifying the data type of input and output