Module libsoxr::api

This is a thin wrapper around the libsoxr C API.

The documentation for this api namespace was copied from soxr.h of libsoxr

API conventions

Buffer lengths (and occupancies) are expressed as the number of contained samples per channel.

Parameter names for buffer lengths have the suffix `len'.

A single-character `i' or 'o' is often used in names to give context as input or output (e.g. ilen, olen).

Structs

soxr_io_spec_t

soxr_quality_spec

soxr_runtime_spec_t

Enums

soxr

soxr_datatype_t
Datatypes supported for I/O to/from the resampler:

Constants

SOXR_16_BITQ

SOXR_20_BITQ

SOXR_24_BITQ

SOXR_28_BITQ

SOXR_32_BITQ

SOXR_COEF_INTERP_AUTO
Auto select coef. interpolation.

SOXR_COEF_INTERP_HIGH
Man. select: more CPU, less memory.

SOXR_COEF_INTERP_LOW
Man. select: less CPU, more memory.

SOXR_DOUBLE_PRECISION
Use D.P. calcs even if precision <= 20.

SOXR_HI_PREC_CLOCK
Increase `irrational' ratio accuracy.

SOXR_HQ
'High quality'.

SOXR_INTERMEDIATE_PHASE

SOXR_LINEAR_PHASE

SOXR_LQ
'Low' 16-bit with larger rolloff.

SOXR_MINIMUM_PHASE

SOXR_MQ
'Medium' 16-bit with medium rolloff.

SOXR_NO_DITHER
Disable the above.

SOXR_QQ
'Quick' cubic interpolation.

SOXR_ROLLOFF_MEDIUM
<= 0.35 dB

SOXR_ROLLOFF_NONE
For Chebyshev bandwidth.

SOXR_ROLLOFF_SMALL
<= 0.01 dB

SOXR_STEEP_FILTER

SOXR_TPDF
Applicable only if otype is INT16.

SOXR_VHQ
'Very high quality'.

SOXR_VR
Variable-rate resampling.

Functions

soxr_clear
Ready for fresh signal, same config.

soxr_create
Create a stream resampler

soxr_delay
Query current delay in output samples

soxr_delete
Free resources.

soxr_engine
Query resampling engine name.

soxr_error
Query error status.

soxr_io_spec
These functions allow setting of the most commonly-used structure parameters, with other parameters being given default values. The default values may then be overridden, directly in the structure, if needed.

soxr_num_clips
Query int. clip counter (for R/W).

soxr_output
Resample and output a block of data. If using an app-supplied input function, it must look and behave like soxr_input_fn_t and be registered with a previously created stream resampler using soxr_set_input_fn then repeatedly call soxr_output.

soxr_process
If not using an app-supplied input function, after creating a stream resampler, repeatedly call soxr_process

soxr_quality_spec
These functions allow setting of the most commonly-used structure parameters, with other parameters being given default values. The default values may then be overridden, directly in the structure, if needed.

soxr_runtime_spec
These functions allow setting of the most commonly-used structure parameters, with other parameters being given default values. The default values may then be overridden, directly in the structure, if needed.

soxr_set_error
Set error of already created soxr_t

soxr_set_input_fn
Set (or reset) an input function.

soxr_set_io_ratio
For variable-rate resampling. See example # 5 for how to create a variable-rate resampler and how to use this function.

soxr_set_num_channels
Set number of channels of already created soxr_t

soxr_version
Query library version: "libsoxr-x.y.z"

Type Definitions

soxr_error_t

soxr_fn_state_t
If needed by the input function. As given to soxr_set_input_fn

soxr_in_t
Either a soxr_cbuf_t or soxr_cbufs_t, depending on itype in soxr_io_spec_t.

soxr_input_fn_t
Function to supply data to be resampled.

soxr_out_t
Either a soxr_buf_t or soxr_bufs_t,depending on otype in soxr_io_spec_t.

soxr_quality_spec_t

soxr_t