Crate libsoxr

libsoxr-rs

This library is a thin wrapper for libsoxr which is a "High quality, one-dimensional sample-rate conversion library".

For direct access to the libsoxr functions, you can use the libsoxr-sys crate.

This wrapper library is licensed the same as libsoxr itself: LGPLv2.

The documentation can be found here.

Install

add the following to your Cargo.toml:

[dependencies.libsoxr]
git = "https://github.com/lrbalt/libsoxr-rs"

and add the crate:

extern crate libsoxr;
use libsoxr::Soxr;

Example

// upscale factor 2, one channel with all the defaults
let s = Soxr::create(1.0, 2.0, 1, None, None, None).unwrap();

// source data, taken from 1-single-block.c of libsoxr examples.
let source: [f32; 48] = [0.0, 1.0, 0.0, -1.0, 0.0, 1.0, 0.0, -1.0, 0.0, 1.0, 0.0, -1.0, 0.0,
                         1.0, 0.0, -1.0, 0.0, 1.0, 0.0, -1.0, 0.0, 1.0, 0.0, -1.0, 0.0, 1.0,
                         0.0, -1.0, 0.0, 1.0, 0.0, -1.0, 0.0, 1.0, 0.0, -1.0, 0.0, 1.0, 0.0,
                         -1.0, 0.0, 1.0, 0.0, -1.0, 0.0, 1.0, 0.0, -1.0];

// create room for 2*48 = 96 samples
let mut target: [f32; 96] = [0.0; 96];

// two runs. last run with None to inform resampler of end-of-input
let result = s.process(Some(&source), &mut target).and_then(|_| {
    s.process::<f32, f32>(None, &mut target[0..]).and_then(|_| {
        for s in target.iter() {
            print!("{:?}\t", s)
        }
        Ok(())
    })
});

Re-exports

pub use crate::datatype::Datatype;

pub use crate::soxr::Soxr;

pub use crate::spec::IOSpec;

pub use crate::spec::QualitySpec;

pub use crate::spec::RuntimeSpec;

Modules

datatype
Encapsulated data types for samples

soxr
Rust API for SOXR.

spec
For specifying the runtime settings of the resampler For specifying the data type of input and output