libsoxr::api::soxr_process

pub unsafe extern fn soxr_process(resampler: soxr_t, buf_in: soxr_in_t, ilen: size_t, idone: *mut size_t, buf_out: soxr_out_t, olen: size_t, odone: *mut size_t) -> soxr_error_t

If not using an app-supplied input function, after creating a stream resampler, repeatedly call soxr_process

resampler - As returned by soxr_create.

Input (to be resampled):

buf_in - Input buffer(s); may be NULL (see below).

ilen - Input buf. length (samples per channel).

idone - To return actual # samples used (<= ilen).

Output (resampled):

buf_out - Output buffer(s).

olen - Output buf. length (samples per channel).

odone - To return actual # samples out (<= olen).

Note that no special meaning is associated with ilen or olen equal to zero. End-of-input (i.e. no data is available nor shall be available) may be indicated by setting `in' to NULL.