Module libsoxr::spec

For specifying the runtime settings of the resampler For specifying the data type of input and output

Structs

IOSpec
Wrapper for soxr_io_spec_t

QualityFlags

QualitySpec
Wrapper for soxr_quality_spec_t

RuntimeSpec
Wrapper for soxr_runtime_spec_t

Enums

QualityRecipe

Constants

HI_PREC_CLOCK

ROLLOFF_MEDIUM

ROLLOFF_NONE

ROLLOFF_SMALL

VR